requirements.txt
fastapi
uvicorn
requests
playwright
pandas
PyPDF2
python-dotenv
google-generativeai
langchain
langgraph
langchain-google-genai
